refactor(agents): rename AgentMonitoring service to Agents, drop tenantId [PLT-103787]
- Rename service AgentMonitoring -> Agents (class AgentService, file agents.ts)
- Flatten module: src/services/agents/monitoring/ -> src/services/agents/agents.ts
- Subpath: @uipath/uipath-typescript/agent-monitoring -> /agents
- Endpoint constant: AGENT_MONITORING_ENDPOINTS -> AGENTS_ENDPOINTS
  (singular AGENT_ENDPOINTS already owned by conversational-agent)
- Remove tenantId param from getNames() - backend no longer requires it
- Drop AGENT_MONITORING_TEST_TENANT_ID env gate from integration tests
- Add OR.Folders.Read to OAuth scope (verified via E2E browser test)
