+++ package.json
@@ -166,14 +166,14 @@
                 "default": "./dist/document-understanding/index.cjs"
             }
         },
-        "./agent-monitoring": {
+        "./agents": {
             "import": {
-                "types": "./dist/agent-monitoring/index.d.ts",
-                "default": "./dist/agent-monitoring/index.mjs"
+                "types": "./dist/agents/index.d.ts",
+                "default": "./dist/agents/index.mjs"
             },
             "require": {
-                "types": "./dist/agent-monitoring/index.d.ts",
-                "default": "./dist/agent-monitoring/index.cjs"
+                "types": "./dist/agents/index.d.ts",
+                "default": "./dist/agents/index.cjs"
             }
         }
     },
