--- src/models/agents/monitoring/index.ts
@@ -1,8 +1,0 @@
-/**
- * Agent Monitoring Types
- *
- * Public types for the Agent Monitoring (real-time monitoring) service.
- */
-
-export * from './monitoring.types';
-export * from './monitoring.models';
