+++ src/utils/constants/endpoints/agents.ts
@@ -1,10 +1,10 @@
 /**
- * Agent Monitoring Service Endpoints
+ * Agent Service Endpoints
  */
 
 import { INSIGHTS_RTM_BASE } from './base';
 
-export const AGENT_MONITORING_ENDPOINTS = {
+export const AGENTS_ENDPOINTS = {
   /** List distinct agent names visible to the caller */
   GET_NAMES: `${INSIGHTS_RTM_BASE}/Agents/names`,
 } as const;
