+++ tests/.env.integration.example
@@ -63,6 +63,3 @@ JOBS_TEST_FOLDER_ID=
 
 # Orchestrator attachment ID (UUID) for attachment getById tests
 ORCHESTRATOR_ATTACHMENT_ID=
-
-# Agent Monitoring tenant ID (UUID) — must match the JWT tenant
-AGENT_MONITORING_TEST_TENANT_ID=
