+++ tests/utils/constants/agents.ts
@@ -1,13 +1,12 @@
 /**
- * Test constants for Agent Monitoring service
+ * Test constants for the Agents service
  */
 
-export const AGENT_MONITORING_TEST_CONSTANTS = {
-  TENANT_ID: '22986e36-8b04-4593-b82f-aae4c14bb2dc',
+export const AGENT_TEST_CONSTANTS = {
   FOLDER_KEY_1: 'c2751834-1f05-4f4e-9cb8-509406f6faac',
   FOLDER_KEY_2: 'f1537e42-50a0-4598-8305-ac95260132cf',
   AGENT_NAME_1: 'ITRReviewerUB',
   AGENT_NAME_2: 'JokeAgent',
   AGENT_NAME_3: 'StoryAgent',
-  ERROR_GENERIC: 'Agent Monitoring API error',
+  ERROR_GENERIC: 'Agents API error',
 } as const;
